src/requirements.txt
fastapi
pydantic
typing
uvicorn
requests
python-jose
redis
passlib
pytz
python-multipart
numpy
pandas
python-magic
pymongo
Pillow
icalendar
pyppeteer
openpyxl
asyncio_redis
aiofiles
cuid2
asyncpg
python-dateutil